+++ screens/partner/table/index.js
@@ -6,36 +6,37 @@ import { useAsyncEffect } from 'use-async-effect'
 import BootstrapTable from 'react-bootstrap-table-next'
 import 'react-bootstrap-table-next/dist/react-bootstrap-table2.min.css'
 import _debounce from 'lodash/debounce'
-
+import _size from 'lodash/size'
 import _reduce from 'lodash/reduce'
-import { getProviderList } from 'services/partner'
+
+import { getPartnerData } from '../../../services/partner'
+// import { getCookie } from '../../../utils/global-util'
+
 import PaginationPanel from '../../../components/extended-table/pagination-panel'
 import PaginationSize from '../../../components/extended-table/pagination-size'
 import PaginationTotal from '../../../components/extended-table/pagination-total'
 import PaginationGoTo from '../../../components/extended-table/pagination-goto'
-import AXASearch from '../../../components/axa-search'
+import Search from '../../../components/extended-table/search'
+// // import Search from '@components/ExtendedTable/Search'
+// import ButtonLoader from '@components/ButtonLoader'
+// import Dropdown from '@components/Dropdown'
 
 import styles from './styles.js'
 import globalStyles from './global-styles'
 
-const LIMIT = 100
 export default function TablePartner({
     columns,
     pageSizes,
     searchOpt,
     textLang,
-    userPosition,
-    sessionToken,
 }) {
     const [partnerData, setPartnerData] = useState({})
-    const [dataSliced, setDataSliced] = useState(null)
     const {
         list: partnerList,
         flattenList: partnerFlatList,
         totalData,
-        offsetFrom,
-        providerTypeList,
-        specialitiesList,
+        searchfieldDt,
+        searchkeyDt,
     } = partnerData
 
     const [sizePerPage, setSizePerPage] = useState(
@@ -44,87 +45,19 @@ export default function TablePartner({
     const [totalPage, setTotalPage] = useState(null)
     const [pageIndex, setPageIndex] = useState(null)
     const [currPage, setCurrPage] = useState(1)
-    const [gotoPage, setGotoPage] = useState()
-    const [searchInitData, setSearchInitData] = useState()
-    const [tableCol, setTableCol] = useState(null)
+    const [goToPage, setGoToPage] = useState('')
+    const [searchObj, setSearchObj] = useState(searchOpt ? searchOpt[0] : null)
 
-    const [selectedOpt, setSelectedOpt] = useState(
-        searchOpt ? searchOpt[0] : null
-    )
-    const [searchObj, setSearchObj] = useState({
-        [searchOpt ? searchOpt[0].id : '']: '',
-    })
-    // const [, setCurrPage] = useState(1)
+    useAsyncEffect(async (isMounted) => {
+        const data = await getDataList({ limit: 10, page: 1 })
 
-    function linkFormatter(cell, row, rowIndex, formatExtraData) {
-        return (
-            <a href={cell} target="_blank" className={'field-link'}>
-                {formatExtraData.linkText}
-            </a>
-        )
-    }
+        if (!isMounted()) return
 
-    function langFormatter(cell) {
-        return textLang[cell] || cell
-    }
-
-    useEffect(() => {
-        //disini nanti ambil dari cms untuk toggle pakai apinya yg mana
-        //check lg perlu toggle atau pisah cms
-        if (columns) {
-            columns.map((col) => {
-                if (col.fieldFormat === 'link') {
-                    col.formatter = linkFormatter
-                }
-
-                if (col.fieldFormat === 'lang') {
-                    col.formatter = langFormatter
-                }
-                return col
-            })
-            setTableCol(columns)
-        }
+        setPartnerData(data)
     }, [])
-
-    useAsyncEffect(
-        async (isMounted) => {
-            if (userPosition) {
-                const data = await getDataList({ limit: 10, page: 1 })
-
-                if (!isMounted()) return
-
-                setPartnerData(data)
-            }
-        },
-        [userPosition]
-    )
-
-    const initSearchList = (opt) => {
-        let searchInitDt = null
-
-        if (opt && opt.id == 'specialities') {
-            searchInitDt =
-                specialitiesList &&
-                specialitiesList.map((dt) => ({
-                    ...dt,
-                    value: dt.label,
-                }))
-        } else if (opt && opt.id == 'providerType') {
-            searchInitDt =
-                providerTypeList &&
-                providerTypeList.map((dt) => ({
-                    ...dt,
-                    label: textLang[dt.label] || dt.label,
-                }))
-        }
-        setSearchInitData(searchInitDt)
-    }
 
     useEffect(() => {
         if (partnerData) {
-            sliceDataPerPage(currPage, sizePerPage)
-            if (!searchInitData) initSearchList(selectedOpt)
-
             const totalPage = Math.ceil(totalData / sizePerPage)
             setTotalPage(totalPage)
             setPageIndex(getPageIndex(totalData, sizePerPage, currPage))
@@ -140,151 +73,87 @@ export default function TablePartner({
     }
 
     const getDataList = async ({
-        limit_page = sizePerPage,
+        limit = sizePerPage,
         page = currPage,
-        searchValue = searchObj[selectedOpt?.id ?? ''] || '',
-        selectedOption = selectedOpt,
+        source = searchObj?.source ?? '/hospitals',
+        param = searchObj?.param ?? '',
+        searchkey = searchObj?.key ?? '',
+        searchfield = searchObj?.field ?? '',
     }) => {
-        let searchVal = searchValue,
-            lat = userPosition.lat,
-            lng = userPosition.long,
-            country = 'Indonesia'
-        if (searchVal) {
-            if (selectedOption.id == 'specialities') {
-                searchVal = ''
-                searchValue &&
-                    searchValue.map((val, index) => {
-                        searchVal += index == 0 ? val.value : `,${val.value}`
-                    })
-            } else if (selectedOption.id == 'location') {
-                lat = searchVal?.lat ?? userPosition.lat
-                lng = searchVal?.lng ?? userPosition.long
-                country = searchVal?.country ?? 'Indonesia'
-            }
-        }
-        setDataSliced(null)
-        const partnerDt = await getProviderList({
-            offset: Math.floor((page - 1) / (LIMIT / limit_page)) * LIMIT,
-            latitude: lat,
-            longitude: lng,
-            providerName: selectedOption.id == 'name' ? searchVal : '',
-            providerType:
-                selectedOption.id == 'providerType' ? searchVal.value : '',
-            specialities: selectedOption.id == 'specialities' ? searchVal : '',
-            country,
+        setPartnerData({})
+        const partnerData = await getPartnerData({
+            limit,
+            page,
+            sort: 'asc',
+            source: param ? source : '/hospitals',
+            param,
+            searchkey,
+            searchfield,
         })
-
-        return partnerDt
-    }
-
-    const onSelectSearch = (opt) => {
-        setSearchObj({ [opt.id]: '' })
-        setSelectedOpt(opt)
-        initSearchList(opt)
+        return partnerData
     }
 
     const onSearch = async (value, opt = null) => {
         setCurrPage(1)
-        setSearchObj({ [opt.id]: value })
-        processSearch(value, opt)
-    }
+        const selOpt = opt || searchObj
 
-    const processSearch = _debounce(async (value, opt) => {
-        const data = await getDataList({
-            page: 1,
-            searchValue: value,
-            selectedOption: opt,
-        })
-        setPartnerData(data)
-    }, 500)
-
-    const sliceDataPerPage = (currentPage, sizePerPage) => {
-        // console.log('partnerFlatList', partnerFlatList)
-        if (partnerFlatList && partnerFlatList.length > 0) {
-            let calcCurrentPage = currentPage
-
-            if (currentPage * sizePerPage > LIMIT) {
-                calcCurrentPage = currentPage % (LIMIT / sizePerPage)
-                /* example
-                    currentPage = 14, sizePerPage = 10
-                    calcCurrentPage = 4
-                */
-                if (calcCurrentPage == 0) {
-                    calcCurrentPage = LIMIT / sizePerPage
-                    /* example
-                    currentPage = 30, sizePerPage = 10
-                    calcCurrentPage = LIMIT / sizePerPage = 10
-
-                    currentPage = 4, sizePerPage = 50
-                    calcCurrentPage = LIMIT / sizePerPage = 2
-                    */
-                }
-            }
-
-            let sliceEnd = calcCurrentPage * sizePerPage
-            if (partnerFlatList.length < calcCurrentPage * sizePerPage) {
-                sliceEnd = partnerFlatList.length
-            }
-            const dtSliced = partnerFlatList.slice(
-                (calcCurrentPage - 1) * sizePerPage,
-                sliceEnd
-            )
-
-            setDataSliced(dtSliced)
-        } else {
-            setDataSliced(null)
-        }
-    }
-
-    const getCurrentOffsetFrom = (currentPage, sizePerPage) => {
-        return Math.floor(currentPage / (LIMIT / sizePerPage)) * LIMIT + 1
-    }
-
-    const onPageChange = async (page) => {
-        //click page navigation
-        const currentPage = page.selected + 1
-        setCurrPage(currentPage)
-        setPageIndex(getPageIndex(totalData, sizePerPage, currentPage))
-        const divCurrPage =
-            currentPage % (LIMIT / sizePerPage) == 0
-                ? currentPage - 1
-                : currentPage
-        const currPageOffsetFrom = getCurrentOffsetFrom(
-            divCurrPage,
-            sizePerPage
-        )
-        if (currPageOffsetFrom !== offsetFrom) {
+        setGoToPage('')
+        if (selOpt && selOpt.master) {
             const data = await getDataList({
-                page: currentPage,
+                page: 1,
+                source: value ? selOpt.dataSource : '/hospitals',
+                param: value,
+                searchfield: '',
+                key: '',
+            })
+
+            setSearchObj({
+                ...selOpt,
+                source: selOpt.dataSource,
+                param: value,
+                field: '',
+                key: '',
             })
             setPartnerData(data)
         } else {
-            sliceDataPerPage(currentPage, sizePerPage)
+            processSearch(value, selOpt)
         }
     }
 
-    const onGoToClick = async (pageNumber) => {
-        if (pageNumber <= totalPage) {
-            setCurrPage(pageNumber)
-            setPageIndex(getPageIndex(totalData, sizePerPage, pageNumber))
-            const divCurrPage =
-                pageNumber % (LIMIT / sizePerPage) == 0
-                    ? pageNumber - 1
-                    : pageNumber
-            const currPageOffsetFrom = getCurrentOffsetFrom(
-                divCurrPage,
-                sizePerPage
-            )
+    const onSelectSearch = (opt) => {
+        setSearchObj(opt)
+    }
 
-            if (currPageOffsetFrom !== offsetFrom) {
-                const data = await getDataList({
-                    page: pageNumber,
-                })
-                setPartnerData(data)
-            } else {
-                sliceDataPerPage(pageNumber, sizePerPage)
-            }
-        }
+    const processSearch = _debounce(async (value, opt) => {
+        // console.log('key:', key, ' field:', field)
+        setSearchObj({
+            ...searchObj,
+            source: '/hospitals',
+            param: '',
+            field: opt.id,
+            key: value,
+        })
+        const data = await getDataList({
+            page: 1,
+            searchfield: opt.id,
+            searchkey: value,
+            param: '',
+        })
+
+        setPartnerData(data)
+    }, 500)
+
+    const onPageChange = async (page) => {
+        //click page navigation
+        // console.log('page', page)
+        // console.log('onPageChange search', searchObj)
+        const currentPage = page.selected + 1
+        setCurrPage(currentPage)
+        setGoToPage('')
+        const data = await getDataList({
+            page: currentPage,
+        })
+        setPartnerData(data)
     }
 
     const onSizePerPageChange = async (sizePerPage) => {
@@ -292,7 +161,7 @@ export default function TablePartner({
         /* SET DARI PAGE 1 LAGI */
         setSizePerPage(sizePerPage)
         setCurrPage(1)
-        setGotoPage('')
+        setGoToPage('')
         const data = await getDataList({
             limit: sizePerPage,
             page: 1,
@@ -300,18 +169,28 @@ export default function TablePartner({
         setPartnerData(data)
     }
 
+    const onGoToClick = async (pageNumber) => {
+        if (pageNumber && pageNumber <= totalPage) {
+            setCurrPage(pageNumber)
+            setGoToPage(pageNumber)
+            const data = await getDataList({
+                page: pageNumber,
+            })
+            setPartnerData(data)
+        }
+    }
+
     const handleNoData = () => {
         if (partnerData.totalData == 0) {
             return (
                 <span style={{ fontWeight: 'normal' }}>
-                    {' '}
                     {textLang['noData']}
                 </span>
             )
         }
         return (
             <span style={{ fontWeight: 'normal' }}>
-                <i> {textLang['loading']}</i>
+                <i>{textLang['loading']}</i>
             </span>
         )
     }
@@ -332,28 +211,23 @@ export default function TablePartner({
                         onClick={onGoToClick}
                         isDesktop={true}
                         textLang={textLang}
-                        value={gotoPage}
+                        value={goToPage}
                     />
-                    <AXASearch
+                    <Search
                         onSelect={onSelectSearch}
                         onSearch={onSearch}
                         options={searchOpt}
-                        initialData={searchInitData}
                         isDesktop={true}
-                        sessionToken={sessionToken}
-                        textLang={textLang}
                     />
                 </div>
-                {tableCol && (
-                    <BootstrapTable
-                        keyField="providerId"
-                        noDataIndication={handleNoData}
-                        data={dataSliced || []}
-                        columns={tableCol}
-                    />
-                )}
-
-                {dataSliced && dataSliced.length > 0 && (
+                <BootstrapTable
+                    keyField="id"
+                    // data={[]}
+                    noDataIndication={handleNoData}
+                    data={partnerFlatList || []}
+                    columns={columns}
+                />
+                {partnerFlatList && partnerFlatList.length > 0 && (
                     <div className={`pagination-wrapper`}>
                         <PaginationTotal
                             from={pageIndex?.startPage ?? 0}